Passing keywords through regular argument splats
This approach uses a flag bit on the final hash object in the regular splat.
If there is a argument splat when calling a method, and no keyword arguments
are specified, and the final argument is a hash with the flag bit set, then
the hash is passed as a keyword splat.  This allows behavior that is more
backwards compatible with Ruby <=2.6. Example:

  def foo(*args)
    @args = args
    bar
  end
  def bar
    baz(*@args)
  end
  def baz(*args, **kw)
    [args, kw]
  end
  foo(a:1)    #=> [[], {a: 1}]
  foo({a: 1}, **{}) #=> [[{a: 1}], {}]

  foo({a: 1}) #=> 2.7: [[], {a: 1}] # and warning
  foo({a: 1}) #=> 3.0: [[{a: 1}], {}]

This fixes argument delegation in the delegate library.

Note that this approach can cause problems:

  def foo(*args)
    args.each do |arg|
      bar(arg, **{})
    end
  end
  def bar(*args, **kw)
    baz(*args) unless kw[:skip]
  end
  def baz(a=1, **kw)
    [h, kw]
  end

Here `foo` is designed to take only positional arguments, not to
pass through keywords.  However, because the last argument to
foo is flagged as keywords, when `baz` is called, the {a: 1} hash
intended to be a positional argument is treated as a keyword
argument.  Note that `baz` could be at a completely different
place in the program, far from the definition of `foo`, which
will make cases like this difficult to debug.